+++ project/codes/buildins/env.c
@@ -6,11 +6,11 @@
 /*   By: vgomes-p <vgomes-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/02/04 16:13:56 by vgomes-p          #+#    #+#             */
-/*   Updated: 2025/02/10 13:37:49 by vgomes-p         ###   ########.fr       */
+/*   Updated: 2025/02/11 15:06:11 by vgomes-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "../includes/minishell.h"
+#include "../../includes/minishell.h"
 
 void	ms_env(t_minishell *shell)
 {
