+++ project/codes/token_and_parse/parse.c
@@ -1,36 +1,13 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   echo.c                                             :+:      :+:    :+:   */
+/*   parse.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: vgomes-p <vgomes-p@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2025/02/04 16:12:45 by vgomes-p          #+#    #+#             */
-/*   Updated: 2025/02/11 15:07:24 by vgomes-p         ###   ########.fr       */
+/*   Created: 2025/02/11 15:05:21 by vgomes-p          #+#    #+#             */
+/*   Updated: 2025/02/11 15:11:14 by vgomes-p         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
-
-void	ms_echo(char **args)
-{
-	int	curr_arg;
-	int	nwline;
-
-	curr_arg = 1;
-	nwline = 1;
-	if (args[1] && lms_strcmp(args[1], "-n") == 0)
-	{
-		nwline = 0;
-		curr_arg++;
-	}
-	while (args[curr_arg])
-	{
-		lms_putstr(args[curr_arg]);
-		if (args[curr_arg + 1])
-			lms_putstr(" ");
-		curr_arg++;
-	}
-	if (nwline)
-		lms_putstr("\n");
-}
