+++ project/codes/token_and_parse/token_utils_1.c
@@ -1,0 +1,67 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   token_utils_1.c                                    :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: vgomes-p <vgomes-p@student.42.fr>          +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2025/02/04 16:18:22 by vgomes-p          #+#    #+#             */
+/*   Updated: 2025/02/11 15:06:46 by vgomes-p         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "../../includes/minishell.h"
+
+bool	is_quotes(char ch)
+{
+	return (ch == '\'' || ch == '\"');
+}
+
+bool	is_operator(char *str)
+{
+	return (lms_strcmp(str, "|") == 0 || lms_strcmp(str, ">") == 0
+		|| lms_strcmp(str, "<") == 0 || lms_strcmp(str, ">>") == 0
+		|| lms_strcmp(str, "<<") == 0);
+}
+
+void	cleanup_tokens(char **tokens, int token_cnt)
+{
+	int	pos;
+
+	pos = 0;
+	while (pos < token_cnt)
+	{
+		free(tokens[pos]);
+		pos++;
+	}
+	free(tokens);
+}
+
+
+void	free_split_array(char **split)
+{
+	int	pos;
+
+	pos = 0;
+	while (split[pos])
+	{
+		free(split[pos]);
+		pos++;
+	}
+	free(split);
+}
+
+int	count_tokens(t_token *tokens)
+{
+	int		count;
+	t_token	*current;
+
+	count = 0;
+	current = tokens;
+	while (current)
+	{
+		count++;
+		current = current->next;
+	}
+	return (count);
+}
